+++ magenpy/GenotypeMatrix.py
@@ -343,7 +343,7 @@ class GenotypeMatrix(object):
     def compute_ld(self,
                    estimator,
                    output_dir,
-                   dtype='int8',
+                   dtype='int16',
                    compressor_name='zstd',
                    compression_level=7,
                    compute_spectral_properties=False,