+++ magenpy/__init__.py
@@ -17,7 +17,7 @@ from .simulation.PhenotypeSimulator import PhenotypeSimulator
 from .utils.data_utils import *
 
 __version__ = '0.1.4'
-__release_date__ = 'October 2024'
+__release_date__ = 'December 2024'
 
 
 config = configparser.ConfigParser()